+++ ffa.js
@@ -167,7 +167,7 @@ var FFA = Tournament.sub('FFA', function (opts, initParent) {
 });
 
 //------------------------------------------------------------------
-// Static helpers and constants
+// Helpers and constants
 //------------------------------------------------------------------
 
 FFA.configure({
@@ -181,6 +181,21 @@ FFA.configure({
     return invalid(np, opts.sizes, opts.advancers, opts.limit);
   }
 });
+
+FFA.prototype.limbo = function (playerId) {
+  // if player reached currentRound, he may be waiting for generation of nextRound
+  var m = $.firstBy(function (m) {
+    return m.p.indexOf(playerId) >= 0 && m.m;
+  }, this.currentRound() || []);
+
+  if (m) {
+    // will he advance to nextRound ?
+    var adv = this.advs[m.id.r - 1];
+    if (Tournament.sorted(m).slice(0, adv).indexOf(playerId) >= 0) {
+      return {s: 1, r: m.id.r + 1}; // TODO: no toString representation for this
+    }
+  }
+};
 
 //------------------------------------------------------------------
 // Expected methods
@@ -209,21 +224,6 @@ FFA.prototype._verify = function (match, score) {
     }
   }
   return null;
-};
-
-FFA.prototype._limbo = function (playerId) {
-  // if player reached currentRound, he may be waiting for generation of nextRound
-  var m = $.firstBy(function (m) {
-    return m.p.indexOf(playerId) >= 0 && m.m;
-  }, this.currentRound() || []);
-
-  if (m) {
-    // will he advance to nextRound ?
-    var adv = this.advs[m.id.r - 1];
-    if (Tournament.sorted(m).slice(0, adv).indexOf(playerId) >= 0) {
-      return {s: 1, r: m.id.r + 1}; // TODO: no toString representation for this
-    }
-  }
 };
 
 FFA.prototype._stats = function (res, m) {